Create books in frontend
Create books in frontend, bug fix in backend, frontend.

+++ Books_Project_Backend/Books_Project/src/main/java/lt/techin/booksproject/dto/BookCreateRequest.java
@@ -20,7 +20,7 @@ public class BookCreateRequest {
 //    )
     private String isbn;
     private String picture;
-    private Long numberOfPages;
+    private Integer numberOfPages;
     @NotNull
     private Long categoryId;
